fix: Decompose provider/index.ts to resolve modularity check
Fixes issue #64 by breaking down the 1128-line src/provider/index.ts file
into smaller, more manageable modules:

- src/provider/portfolioUtils.ts: Portfolio calculation utilities (50 lines)
- src/provider/orderManager.ts: Order generation and management (264 lines)
- src/provider/marketData.ts: Market data and exchange functions (212 lines)
- src/provider/accountUtils.ts: Account configuration utilities (69 lines)
- src/provider/index.ts: Main provider logic (541 lines)

This modular structure improves code maintainability and ensures all files
are within the 1000-line modularity check limit. All functions are properly
exported and imported, maintaining the same functionality while improving
code organization.
