+++ src/provider/index.ts
@@ -18,68 +18,18 @@ import { expenseTracker, ExpenseRecord } from '../expenseTracker';
 import { ProfitCalculator } from '../profitCalculator';
 import { dailyAggregator } from '../dailyAggregator';
 
+// Import modularized functions
+import { filterFrozenAssets, calculateAvailablePortfolioValue, calculatePortfolioShares } from './portfolioUtils';
+import { generateOrders, generateOrdersSequential, generateOrder } from './orderManager';
+import { isExchangeOpenNow, getLastPrice, getInstruments, getLastPrices } from './marketData';
+import { getAccountConfigById, getAccountId } from './accountUtils';
+
 (global as any).INSTRUMENTS = [];
 (global as any).POSITIONS = [];
 (global as any).LAST_PRICES = [];
 
 const debugProvider = debug('bot').extend('provider');
 
-// Helper function to get account config by ID
-const getAccountConfigById = (accountId: string) => {
-  const account = configLoader.getAccountById(accountId);
-  if (!account) {
-    throw new Error(`Account with id '${accountId}' not found in CONFIG.json`);
-  }
-  return account;
-};
-
-/**
- * Filters out frozen (blocked) assets from the wallet
- * @param wallet - array of portfolio positions
- * @returns wallet with only available (non-frozen) assets
- */
-const filterFrozenAssets = (wallet: Wallet): Wallet => {
-  return wallet.filter(position => !position.blocked);
-};
-
-/**
- * Calculates the total value of available (non-frozen) assets
- * @param wallet - array of portfolio positions
- * @returns total value of available assets in RUB
- */
-const calculateAvailablePortfolioValue = (wallet: Wallet): number => {
-  const availableWallet = filterFrozenAssets(wallet);
-  return _.sumBy(availableWallet, 'totalPriceNumber');
-};
-
-/**
- * Рассчитывает доли каждого инструмента в портфеле
- * @param wallet - массив позиций портфеля
- * @param includeBlocked - включать заблокированные позиции в расчет (по умолчанию false)
- * @returns объект с тикерами и их долями в процентах
- */
-const calculatePortfolioShares = (wallet: Wallet, includeBlocked: boolean = false): Record<string, number> => {
-  // Исключаем валюты (позиции где base === quote)
-  let securities = wallet.filter(p => p.base !== p.quote);
-
-  // Filter out frozen assets unless explicitly requested to include them
-  if (!includeBlocked) {
-    securities = filterFrozenAssets(securities);
-  }
-
-  const totalValue = _.sumBy(securities, 'totalPriceNumber');
-
-  if (totalValue <= 0) return {};
-
-  const shares: Record<string, number> = {};
-  for (const position of securities) {
-    if (position.base && position.totalPriceNumber) {
-      const ticker = normalizeTicker(position.base) || position.base;
-      shares[ticker] = (position.totalPriceNumber / totalValue) * 100;
-    }
-  }
-  return shares;
-};
 
 let ACCOUNT_ID: string;
 
@@ -103,340 +53,10 @@ export const provider = async (options?: { runOnce?: boolean; accountId?: string
   // Use account_id from account configuration
   ACCOUNT_ID = await getAccountId(accountConfig.account_id, users);
   await getInstruments(instruments, accountConfig);
-  await getPositionsCycle(options, accountConfig, { orders, operations, marketData, instruments });
+  await getPositionsCycle(options, accountConfig, { orders, operations, marketData, instruments, accountId: ACCOUNT_ID });
 };
 
-export const generateOrders = async (wallet: Wallet, accountConfig?: any, sdkObjects?: any) => {
-  debugProvider('generateOrders');
-  for (const position of wallet) {
-    await generateOrder(position, accountConfig, sdkObjects);
-  }
-};
 
-/**
- * Generates orders with sequential execution groups for buy_requires_total_marginal_sell feature
- * Ensures sell orders complete before non-margin buy orders are executed
- */
-export const generateOrdersSequential = async (
-  sellsFirst: Position[],
-  buysNonMarginFirst: Position[],
-  remainingOrders: Position[],
-  accountConfig?: any,
-  sdkObjects?: any
-) => {
-  debugProvider('generateOrdersSequential - executing in phases for buy_requires_total_marginal_sell');
-  
-  // Phase 1: Execute sell orders first and wait for completion
-  if (sellsFirst.length > 0) {
-    debugProvider(`🔄 PHASE 1: Executing ${sellsFirst.length} sell orders to raise funds`);
-    for (const position of sellsFirst) {
-      debugProvider(`💰 Executing sell order: ${Math.abs(position.toBuyLots || 0)} lots of ${position.base}`);
-      await generateOrder(position, accountConfig, sdkObjects);
-    }
-    
-    // Additional wait time for sell orders to complete and funds to become available
-    debugProvider('⏳ Waiting for sell orders to complete and funds to be available...');
-    await sleep(5000); // Wait 5 seconds for market orders to complete
-  }
-  
-  // Phase 2: Execute non-margin buy orders (TMON etc.)
-  if (buysNonMarginFirst.length > 0) {
-    debugProvider(`🔄 PHASE 2: Executing ${buysNonMarginFirst.length} non-margin buy orders with raised funds`);
-    for (const position of buysNonMarginFirst) {
-      debugProvider(`💰 Executing non-margin buy order: ${Math.abs(position.toBuyLots || 0)} lots of ${position.base}`);
-      await generateOrder(position, accountConfig, sdkObjects);
-    }
-  }
-  
-  // Phase 3: Execute remaining orders normally
-  if (remainingOrders.length > 0) {
-    debugProvider(`🔄 PHASE 3: Executing ${remainingOrders.length} remaining orders`);
-    for (const position of remainingOrders) {
-      await generateOrder(position, accountConfig, sdkObjects);
-    }
-  }
-  
-  debugProvider('✅ Sequential order execution completed');
-};
-
-export const generateOrder = async (position: Position, accountConfig?: any, sdkObjects?: any): Promise<OrderResult | number | boolean> => {
-  debugProvider('generateOrder');
-  debugProvider('position', position);
-
-  if (position.base === 'RUB') {
-    debugProvider('If position is RUB, do nothing');
-    return false;
-  }
-
-  // Log order details
-  debugProvider(`💰 About to place ${position.toBuyLots! > 0 ? 'BUY' : 'SELL'} order: ${Math.abs(position.toBuyLots || 0)} lots × ${position.lotPriceNumber || 0} = ${Math.abs((position.toBuyLots || 0) * (position.lotPriceNumber || 0)).toFixed(2)} RUB for ${position.base}`);
-
-  debugProvider('Position is not currency');
-
-  debugProvider('position.toBuyLots', position.toBuyLots);
-
-  if (!position.toBuyLots || !isFinite(position.toBuyLots)) {
-    debugProvider('toBuyLots is NaN/Infinity/undefined. Skipping position.');
-    const result: OrderExecutionResult = {
-      status: 'skipped',
-      errorMessage: 'toBuyLots is NaN/Infinity/undefined'
-    };
-    return result;
-  }
-
-  if ((-1 < position.toBuyLots) && (position.toBuyLots < 1)) {
-    debugProvider('Order less than 1 lot. Not worth executing.');
-    const result: OrderExecutionResult = {
-      status: 'skipped',
-      errorMessage: 'Order less than 1 lot'
-    };
-    return result;
-  }
-
-  debugProvider('Position is greater than or equal to 1 lot');
-
-  const direction = position.toBuyLots >= 1 ? OrderDirection.ORDER_DIRECTION_BUY : OrderDirection.ORDER_DIRECTION_SELL;
-  debugProvider('direction', direction);
-
-  // Calculate order value and check for confirmation requirements
-  const quantityLots = Math.floor(Math.abs(position.toBuyLots || 0));
-  const orderValueRub = quantityLots * (position.lotPriceNumber || 0);
-
-  // Check if confirmation is required for large orders
-  if (accountConfig?.analysis?.openrouter?.requireConfirmationForLargeOrders &&
-      accountConfig?.confirmationThresholdRub !== undefined &&
-      orderValueRub > accountConfig.confirmationThresholdRub) {
-
-    debugProvider(`🔔 Order requires confirmation: ${orderValueRub.toFixed(2)} RUB > ${accountConfig.confirmationThresholdRub} RUB threshold`);
-
-    // Create order details for confirmation request
-    const orderDetails: OrderDetails = {
-      ticker: position.base || '',
-      quantity: quantityLots,
-      direction: direction === OrderDirection.ORDER_DIRECTION_BUY ? 'BUY' : 'SELL',
-      valueRub: orderValueRub,
-      lotSize: position.lotSize || 1,
-      pricePerLot: position.lotPriceNumber || 0,
-      figi: position.figi || '',
-      orderId: uniqid()
-    };
-
-    const confirmationRequest: OrderConfirmationRequest = {
-      status: 'needs_confirmation',
-      orderDetails,
-      thresholdInfo: {
-        configuredThreshold: accountConfig.confirmationThresholdRub,
-        actualOrderValue: orderValueRub
-      },
-      timestamp: new Date().toISOString()
-    };
-
-    console.log(`\n🔔 CONFIRMATION REQUIRED for ${orderDetails.direction} order:`);
-    console.log(`   Ticker: ${orderDetails.ticker}`);
-    console.log(`   Quantity: ${orderDetails.quantity} lots`);
-    console.log(`   Value: ${orderValueRub.toFixed(2)} RUB`);
-    console.log(`   Threshold: ${accountConfig.confirmationThresholdRub} RUB`);
-    console.log(`   Order ID: ${orderDetails.orderId}`);
-    console.log(`   Awaiting human approval...`);
-
-    return confirmationRequest;
-  }
-
-  // for (const i of _.range(position.toBuyLots)) {
-  //   // Idea to create single-lot orders to ensure they always execute completely, not partially.
-  //   // May have complications with:
-  //   // - number of allowed API requests, then need to implement queue.
-  //   // - minimum order may be more than one lot
-  //   debugProvider(`Creating single-lot order #${i} of ${_.range(position.toBuyLots).length}`);
-  //   const order = {
-  //     accountId: ACCOUNT_ID,
-  //     figi: position.figi,
-  //     quantity: 1,
-  //     // price: { units: 40, nano: 0 },
-  //     direction,
-  //     orderType: OrderType.ORDER_TYPE_MARKET,
-  //     orderId: uniqid(),
-  //   };
-  //   debugProvider('Sending order', order);
-
-  //   try {
-  //     const setOrder = await orders.postOrder(order);
-  //     debugProvider('Successfully placed order', setOrder);
-  //   } catch (err) {
-  //     debugProvider('Error placing order');
-  //     debugProvider(err);
-  //     console.trace(err);
-  //   }
-  //   await sleep(1000);
-  // }
-
-  // Continue with normal order execution
-  debugProvider('position', position);
-
-  debugProvider('Creating market order');
-
-  if (quantityLots < 1) {
-    debugProvider('Number of lots after rounding < 1. Skipping order.');
-    const result: OrderExecutionResult = {
-      status: 'skipped',
-      errorMessage: 'Number of lots after rounding < 1'
-    };
-    return result;
-  }
-
-  if (!position.figi) {
-    debugProvider('Position missing figi. Skipping order.');
-    const result: OrderExecutionResult = {
-      status: 'skipped',
-      errorMessage: 'Position missing figi'
-    };
-    return result;
-  }
-
-  const order = {
-    accountId: ACCOUNT_ID,
-    figi: position.figi,
-    quantity: quantityLots, // Number of lots must be integer
-    // price: { units: 40, nano: 0 },
-    direction,
-    orderType: OrderType.ORDER_TYPE_MARKET,
-    orderId: uniqid(),
-  };
-  debugProvider('Sending market order', order);
-
-  const { orders } = sdkObjects || {};
-  if (!orders) {
-    debugProvider('orders SDK object not provided, skipping order');
-    const result: OrderExecutionResult = {
-      status: 'skipped',
-      errorMessage: 'orders SDK object not provided'
-    };
-    return result;
-  }
-
-  try {
-    const setOrder = await orders.postOrder(order);
-    debugProvider('Successfully placed order', setOrder);
-
-    // Track commission expense
-    let commission = 0;
-    if (setOrder) {
-      // Calculate commission (standard Tinkoff commission is 0.3% for market orders)
-      const orderAmount = quantityLots * (position.lotSize || 1) * (position.priceNumber || 0);
-      commission = orderAmount * 0.003; // 0.3% commission
-
-      const expenseRecord: ExpenseRecord = {
-        orderId: order.orderId,
-        ticker: position.base || '',
-        orderType: direction === OrderDirection.ORDER_DIRECTION_BUY ? 'BUY' : 'SELL',
-        lots: quantityLots,
-        amountRub: orderAmount,
-        commission: commission,
-        timestamp: new Date()
-      };
-
-      expenseTracker.addExpense(expenseRecord);
-      debugProvider(`Tracked commission: ${commission.toFixed(2)} RUB for ${position.base}`);
-    }
-
-    await sleep(accountConfig?.sleep_between_orders || 1000);
-
-    const result: OrderExecutionResult = {
-      status: 'executed',
-      orderDetails: {
-        ticker: position.base || '',
-        quantity: quantityLots,
-        direction: direction === OrderDirection.ORDER_DIRECTION_BUY ? 'BUY' : 'SELL',
-        valueRub: orderValueRub,
-        lotSize: position.lotSize || 1,
-        pricePerLot: position.lotPriceNumber || 0,
-        figi: position.figi || '',
-        orderId: order.orderId
-      },
-      commission
-    };
-    return result;
-
-  } catch (err) {
-    debugProvider('Error placing order');
-    debugProvider(err);
-    // console.trace(err);
-
-    await sleep(accountConfig?.sleep_between_orders || 1000);
-
-    const result: OrderExecutionResult = {
-      status: 'error',
-      errorMessage: err instanceof Error ? err.message : 'Unknown error placing order',
-      orderDetails: {
-        ticker: position.base || '',
-        quantity: quantityLots,
-        direction: direction === OrderDirection.ORDER_DIRECTION_BUY ? 'BUY' : 'SELL',
-        valueRub: orderValueRub,
-        lotSize: position.lotSize || 1,
-        pricePerLot: position.lotPriceNumber || 0,
-        figi: position.figi || '',
-        orderId: order.orderId
-      }
-    };
-    return result;
-  }
-};
-
-export const getAccountId = async (type: any, users?: any) => {
-  // Поддержка индекса: '3' или 'INDEX:3'
-  const indexMatch = typeof type === 'string' && type.startsWith('INDEX:')
-    ? Number(type.split(':')[1])
-    : (typeof type === 'string' && /^\d+$/.test(type) ? Number(type) : null);
-
-  // If specific string id was passed, return as is
-  if (indexMatch === null && type !== 'ISS' && type !== 'BROKER') {
-    debugProvider('Passed ACCOUNT_ID (as string id)', type);
-    return type;
-  }
-
-  debugProvider('Getting accounts list');
-  let accountsResponse: any;
-  try {
-    accountsResponse = await users.getAccounts({});
-  } catch (err) {
-    debugProvider('Error getting accounts list');
-    debugProvider(err);
-  }
-  debugProvider('accountsResponse', accountsResponse);
-
-  // Support different response formats: { accounts: [...] } or direct array
-  const accounts: any[] = Array.isArray(accountsResponse)
-    ? accountsResponse
-    : (accountsResponse?.accounts || []);
-
-  // Selection by index
-  if (indexMatch !== null) {
-    const byIndex = accounts[indexMatch];
-    const byIndexId = byIndex?.id || byIndex?.accountId || byIndex?.account_id;
-    debugProvider('Selected account by index', byIndex);
-    if (!byIndexId) {
-      throw new Error(`Could not determine ACCOUNT_ID by index ${indexMatch}.`);
-    }
-    return byIndexId;
-  }
-
-  // Selection by type
-  if (type === 'ISS' || type === 'BROKER') {
-    // 1 — brokerage, 2 — IIS (by API v2 enum)
-    const desiredType = type === 'ISS' ? 2 : 1;
-    const account = _.find(accounts, { type: desiredType });
-    debugProvider('Found account by type', account);
-    const accountId = account?.id || account?.accountId || account?.account_id;
-    if (!accountId) {
-      throw new Error('Could not determine ACCOUNT_ID by type. Check token access to the required account.');
-    }
-    return accountId;
-  }
-
-  // Fallback: return as is
-  debugProvider('Passed ACCOUNT_ID (as string id fallback)', type);
-  return type;
-};
 
 export const getPositionsCycle = async (options?: { runOnce?: boolean }, accountConfig?: any, sdkObjects?: any) => {
   return await new Promise<void>((resolve) => {
@@ -919,210 +539,3 @@ export const getPositionsCycle = async (options?: { runOnce?: boolean }, account
   });
 };
 
-// Преобразование типов времени из ответа API к Date
-const toDate = (t: any): Date | null => {
-  if (!t) return null;
-  if (t instanceof Date) return t;
-  if (typeof t === 'string' || typeof t === 'number') return new Date(t);
-  if (typeof t === 'object') {
-    const seconds = (t.seconds !== undefined ? Number(t.seconds) : (t.sec !== undefined ? Number(t.sec) : undefined));
-    const nanos = (t.nanos !== undefined ? Number(t.nanos) : (t.nano !== undefined ? Number(t.nano) : 0));
-    if (seconds !== undefined) {
-      return new Date(seconds * 1000 + Math.floor(nanos / 1e6));
-    }
-  }
-  return null;
-};
-
-// Проверяет, открыта ли указанная биржа прямо сейчас по расписанию торгов
-export const isExchangeOpenNow = async (exchange: string = 'MOEX', instruments?: any): Promise<boolean> => {
-  try {
-    const now = new Date();
-    const from = new Date(now); // Use current time as 'from' parameter
-    const to = new Date(now);
-    to.setDate(to.getDate() + 1); // Get schedule until tomorrow
-
-    debugProvider(`Checking trading schedule for ${exchange}. Current time: ${now.toISOString()}`);
-    debugProvider(`Request params: from=${from.toISOString()}, to=${to.toISOString()}`);
-
-    if (!instruments) {
-      debugProvider('instruments SDK object not provided, assuming exchange is open');
-      return true;
-    }
-
-    const schedules: any = await instruments.tradingSchedules({
-      exchange,
-      from,
-      to,
-    });
-
-    debugProvider('Trading schedules response:', JSON.stringify(schedules, null, 2));
-
-    const exchanges = schedules?.exchanges || schedules?.exchangesList || [];
-    const first = exchanges[0];
-    const days = first?.days || first?.daysList || [];
-
-    debugProvider(`Found ${days.length} trading days in schedule`);
-
-    // Ищем интервал(ы) сегодняшнего дня и проверяем попадание now
-    for (const day of days) {
-      debugProvider('Processing day:', JSON.stringify(day, null, 2));
-
-      // В некоторых обёртках может быть date как строка/Date — но для надёжности сверяем по границам
-      if (day?.isTradingDay === false) {
-        debugProvider('Day is not a trading day, skipping');
-        continue;
-      }
-
-      const start = toDate(day?.startTime || day?.start_time);
-      const end = toDate(day?.endTime || day?.end_time);
-      const eveningStart = toDate(day?.eveningStartTime || day?.evening_start_time);
-      const eveningEnd = toDate(day?.eveningEndTime || day?.evening_end_time);
-
-      debugProvider(`Session times: start=${start?.toISOString()}, end=${end?.toISOString()}`);
-      debugProvider(`Evening session: start=${eveningStart?.toISOString()}, end=${eveningEnd?.toISOString()}`);
-
-      // Основная сессия
-      if (start && end && now >= start && now <= end) {
-        debugProvider('Current time is within main trading session');
-        return true;
-      }
-      // Вечерняя сессия (если есть)
-      if (eveningStart && eveningEnd && now >= eveningStart && now <= eveningEnd) {
-        debugProvider('Current time is within evening trading session');
-        return true;
-      }
-    }
-
-    debugProvider('Current time is outside all trading sessions');
-    return false;
-  } catch (err) {
-    // In case of errors, don't block bot operation
-    debugProvider('Error requesting trading schedule', err);
-    return true;
-  }
-};
-
-export const getLastPrice = async (figi: any, marketData?: any, accountConfig?: any) => {
-  debugProvider('Getting last price');
-  if (!marketData) {
-    debugProvider('marketData SDK object not provided');
-    return null;
-  }
-  let lastPriceResult;
-  try {
-    lastPriceResult = await marketData.getLastPrices({
-      figi: [figi],
-    });
-    debugProvider('lastPriceResult', lastPriceResult);
-  } catch (err) {
-    debugProvider(err);
-  }
-
-  const lastPrice = lastPriceResult?.lastPrices?.[0]?.price;
-  debugProvider('lastPrice', lastPrice);
-  await sleep(accountConfig?.sleep_between_orders || 1000);
-  return lastPrice;
-};
-
-export const getInstruments = async (instruments?: any, accountConfig?: any) => {
-  if (!instruments) {
-    debugProvider('instruments SDK object not provided');
-    return;
-  }
-
-  debugProvider('Getting shares list');
-  let sharesResult;
-  try {
-    sharesResult = await instruments.shares({
-      // instrumentStatus: InstrumentStatus.INSTRUMENT_STATUS_BASE,
-    });
-  } catch (err) {
-    debugProvider(err);
-  }
-  const shares = sharesResult?.instruments;
-  debugProvider('shares count', shares?.length);
-  (global as any).INSTRUMENTS = _.union(shares, (global as any).INSTRUMENTS);
-  await sleep(accountConfig?.sleep_between_orders || 1000);
-
-  debugProvider('Getting ETFs list');
-  let etfsResult;
-  try {
-    etfsResult = await instruments.etfs({
-      // instrumentStatus: InstrumentStatus.INSTRUMENT_STATUS_BASE,
-    });
-  } catch (err) {
-    debugProvider(err);
-  }
-  const etfs = etfsResult?.instruments;
-  debugProvider('etfs count', etfs?.length);
-  (global as any).INSTRUMENTS = _.union(etfs, (global as any).INSTRUMENTS);
-  await sleep(accountConfig?.sleep_between_orders || 1000);
-
-  debugProvider('Getting bonds list');
-  let bondsResult;
-  try {
-    bondsResult = await instruments.bonds({
-      // instrumentStatus: InstrumentStatus.INSTRUMENT_STATUS_BASE,
-    });
-  } catch (err) {
-    debugProvider(err);
-  }
-  const bonds = bondsResult?.instruments;
-  debugProvider('bonds count', bonds?.length);
-  (global as any).INSTRUMENTS = _.union(bonds, (global as any).INSTRUMENTS);
-  await sleep(accountConfig?.sleep_between_orders || 1000);
-
-  debugProvider('Getting currencies list');
-  let currenciesResult;
-  try {
-    currenciesResult = await instruments.currencies({
-      // instrumentStatus: InstrumentStatus.INSTRUMENT_STATUS_BASE,
-    });
-  } catch (err) {
-    debugProvider(err);
-  }
-  const currencies = currenciesResult?.instruments;
-  debugProvider('currencies count', currencies?.length);
-  (global as any).INSTRUMENTS = _.union(currencies, (global as any).INSTRUMENTS);
-  await sleep(accountConfig?.sleep_between_orders || 1000);
-
-  debugProvider('Getting futures list');
-  let futuresResult;
-  try {
-    futuresResult = await instruments.futures({
-      // instrumentStatus: InstrumentStatus.INSTRUMENT_STATUS_BASE,
-    });
-  } catch (err) {
-    debugProvider(err);
-  }
-  const futures = futuresResult?.instruments;
-  debugProvider('futures count', futures?.length);
-  (global as any).INSTRUMENTS = _.union(futures, (global as any).INSTRUMENTS);
-  await sleep(accountConfig?.sleep_between_orders || 1000);
-
-  debugProvider('=========================');
-};
-
-export const getLastPrices = async (marketData?: any) => {
-  if (!marketData) {
-    debugProvider('marketData SDK object not provided');
-    return;
-  }
-  const lastPrices = (await marketData.getLastPrices({
-    figi: [],
-  }))?.lastPrices;
-  debugProvider('lastPrices', JSON.stringify(lastPrices, null, 2));
-  const lastPricesFormatted = _.map(lastPrices, (item) => {
-    if (item.price) {
-      const priceNumber = convertTinkoffNumberToNumber(item.price);
-      (item as any).price = priceNumber;
-      debugProvider('fffff', priceNumber);
-    }
-    return item;
-  });
-  debugProvider('lastPricesFormatted', JSON.stringify(lastPricesFormatted, null, 2));
-  (global as any).LAST_PRICES = lastPricesFormatted;
-
-  writeFile(lastPricesFormatted, 'lastPrices');
-};
